Separate Publishing API pact tests to separate files
I decided that one file was too big (currently nearly 2000 lines), so
started a process to create individual files for each publishing API
tested endpoint.

It also breaks out the non-pact tests to be into a separate file
(test/publishing_api_test.rb) which covers the unit tests.
